Split specs into separate files
